requirements.txt
fastapi
uvicorn
newspaper3k
requests
transformers
torch
nltk
streamlit
python-dotenv